refactor: make message property setters fallible
Support for `content-type` and metadata/headers across different message
queues and services is not uniform. For example, RabbitMQ has built-in
support for `content-type`, but NATS.io does not.

Make setters of these properties on the message resource fallible to
account for this difference.

Note, that this is only a partial solution to the bigger issue:
What's the correct behavior on the host's part when content-type and/or
metadata is set in the guest message, but not supported by the
underlying implementation?

Signed-off-by: Roman Volosatovs <rvolosatovs@riseup.net>